pyproject.toml
[tool.poetry]
name = "interface-tools"
version = "0.1.0"
description = ""
authors = ["Equinor"]

packages = [
    { include = "interface_tools" }
]

[tool.poetry.dependencies]
python = "^3.7,<3.9"
joblib = "^0.16.0"
azure-storage-file-share = "^12.1.2"
pylint = "^2.5.3"
azureml-sdk = "^1.17.0"
matplotlib = "^3.3.2"
pandas = "^1.1.3"
kaleido = "^0.0.3"
plotly = "^4.11.0"
psutil = "^5.7.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyter = "^1.0.0"
isort = "^5.4.2"
mypy = "^0.782"
flake8 = "^3.8.3"
pre-commit = "^2.7.1"
pydocstyle = "^5.1.1"
pytest-cov = "^2.10.1"
black = "^20.8b1"
openpyxl = "^3.0.5"
sphinx = "^3.2.1"
sphinx-rtd-theme = "^0.5.0"
pytest-env = "^0.6.2"
graphviz = "^0.15"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"